refacto geojson code (#111)
